refactor(cosmetics): modularize UI components and standardize design language
This commit refactors the cosmetics feature module by extracting reusable UI elements into a dedicated `components` package. This organization reduces code duplication across feature screens, improves maintainability, and ensures a consistent visual identity for the "Atelier" design language.

- **`ui/components/`**:
    - Created a comprehensive library of reusable composables including `CosmeticProductCard`, `InventoryProductCard`, `VanityCategoryCard`, and `SummaryStatCard`.
    - Implemented specialized ranking and efficiency rows: `UsageRankingRow` and `ValueEfficiencyRow`.
    - Modularized the cosmetic detail view into logical sections within `DetailSections.kt` (e.g., `ClinicalSafetySection`, `IngredientAnalysisSection`, `SustainabilitySection`).
    - Standardized interactive elements like `TaxonomyDropdown` and `CategoryIconItem`.
- **Screen Refactoring**:
    - Updated `CosmeticsScreen.kt`, `CosmeticDetailScreen.kt`, `VanityLandingScreen.kt`, and `InventoryManagementScreen.kt` to utilize the new modular components.
    - Removed redundant private composables from screen files to improve readability and focus on state management.
- **`ValidateItemScreen.kt`**:
    - Extracted the item validation logic from `StitchProductBuilder.kt` into a standalone screen component to better manage the product discovery flow.
- **Localization & Resources**:
    - Integrated `stringResource` calls across all components to support better localization and moved hardcoded strings to the `strings.xml` resource file.
- **Design & UX**:
    - Standardized typography using `FontFamily.Serif` for headings and consistent corner radii (12dp to 32dp) for cards and containers.
    - Enhanced visual feedback with improved `LinearProgressIndicator` implementations for stock status and usage metrics.
    - Added `@Preview` functions to all new components to facilitate UI development in the Android Studio Layout Editor.
