+++ applications/kocolor/features/cosmetics/src/main/java/com/zoewave/probase/kocolor/features/cosmetics/ui/CosmeticsScreen.kt
@@ -1,15 +1,9 @@
 package com.zoewave.probase.kocolor.features.cosmetics.ui
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.automirrored.filled.Sort
@@ -18,33 +12,16 @@ import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.res.stringResource
-import coil.compose.AsyncImage
-import com.zoewave.probase.features.graphics.colorpicker.ui.ColorPickerDialog
-import com.zoewave.probase.features.graphics.colorpicker.util.isColorDark
-import com.zoewave.probase.features.graphics.colorpicker.util.parseColor
-import com.zoewave.probase.features.graphics.colorpicker.util.toHex
 import com.zoewave.probase.kocolor.features.cosmetics.R
+import com.zoewave.probase.kocolor.features.cosmetics.ui.components.CosmeticProductCard
+import com.zoewave.probase.kocolor.features.cosmetics.ui.components.GroupSectionCard
+import com.zoewave.probase.kocolor.features.cosmetics.ui.components.SubCategoryCard
 import com.zoewave.probase.kocolor.model.*
-
-@Preview(showBackground = true)
-@Composable
-private fun CosmeticsUiRoutePreview() {
-    MaterialTheme {
-        CosmeticsUiRoute(
-            uiState = CosmeticsUiState(),
-            onEvent = {},
-            navTo = {}
-        )
-    }
-}
 
 @Composable
 fun CosmeticsUiRoute(
@@ -57,6 +34,18 @@ fun CosmeticsUiRoute(
         onEvent = onEvent,
         navTo = navTo
     )
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun CosmeticsScreenPreview() {
+    MaterialTheme {
+        CosmeticsScreen(
+            uiState = CosmeticsUiState(),
+            onEvent = {},
+            navTo = {}
+        )
+    }
 }
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -187,141 +176,6 @@ fun CosmeticsScreen(
                         }
                         item { Spacer(modifier = Modifier.height(100.dp)) }
                     }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun GroupSectionCard(
-    title: String,
-    itemCount: Int,
-    isExpanded: Boolean,
-    onToggle: (Boolean) -> Unit
-) {
-    ElevatedCard(
-        modifier = Modifier.fillMaxWidth(),
-        shape = RoundedCornerShape(24.dp),
-        onClick = { onToggle(!isExpanded) }
-    ) {
-        Row(
-            modifier = Modifier.padding(20.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Box(
-                modifier = Modifier
-                    .size(48.dp)
-                    .clip(CircleShape)
-                    .background(MaterialTheme.colorScheme.primary.copy(alpha = 0.1f)),
-                contentAlignment = Alignment.Center
-            ) {
-                Icon(Icons.Default.Category, null, tint = MaterialTheme.colorScheme.primary)
-            }
-            
-            Spacer(modifier = Modifier.width(16.dp))
-            
-            Column(modifier = Modifier.weight(1f)) {
-                Text(text = title, style = MaterialTheme.typography.titleLarge, fontWeight = FontWeight.Black)
-                Text(text = stringResource(R.string.applications_kocolor_features_cosmetics_item_count_format, itemCount), style = MaterialTheme.typography.labelMedium, color = MaterialTheme.colorScheme.onSurfaceVariant)
-            }
-            
-            Icon(
-                imageVector = if (isExpanded) Icons.Default.KeyboardArrowUp else Icons.Default.KeyboardArrowDown,
-                contentDescription = null
-            )
-        }
-    }
-}
-
-@Composable
-fun SubCategoryCard(
-    micro: MicroCategory,
-    itemCount: Int,
-    modifier: Modifier = Modifier
-) {
-    Card(
-        modifier = modifier.fillMaxWidth(),
-        shape = RoundedCornerShape(16.dp),
-        colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.3f))
-    ) {
-        Row(
-            modifier = Modifier.padding(16.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween
-        ) {
-            Text(
-                text = micro.displayName,
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.Bold,
-                color = MaterialTheme.colorScheme.secondary
-            )
-            Badge(containerColor = MaterialTheme.colorScheme.secondary) {
-                Text(text = itemCount.toString(), modifier = Modifier.padding(horizontal = 4.dp))
-            }
-        }
-    }
-}
-
-@Composable
-fun CosmeticProductCard(
-    uiState: CosmeticItem,
-    onEvent: (CosmeticsEvent) -> Unit,
-    navTo: (KoColorRoute) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val cardColor = uiState.colorHex?.let { parseColor(it) } ?: MaterialTheme.colorScheme.surface
-    val contentColor = if (uiState.colorHex != null && isColorDark(cardColor)) Color.White else Color.Black
-
-    ElevatedCard(
-        modifier = modifier
-            .fillMaxWidth()
-            .clickable { navTo(KoColorRoute.CosmeticDetail(uiState.id)) },
-        shape = RoundedCornerShape(20.dp),
-        colors = CardDefaults.elevatedCardColors(containerColor = cardColor, contentColor = contentColor)
-    ) {
-        Column(modifier = Modifier.padding(16.dp)) {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                // Product Image or Swatch
-                Box(
-                    modifier = Modifier
-                        .size(64.dp)
-                        .clip(RoundedCornerShape(12.dp))
-                        .background(contentColor.copy(alpha = 0.1f)),
-                    contentAlignment = Alignment.Center
-                ) {
-                    if (uiState.imageUrl != null) {
-                        AsyncImage(model = uiState.imageUrl, contentDescription = null, modifier = Modifier.fillMaxSize(), contentScale = ContentScale.Crop)
-                    } else {
-                        Icon(Icons.Default.AutoAwesome, null, tint = contentColor.copy(alpha = 0.5f))
-                    }
-                }
-                
-                Spacer(modifier = Modifier.width(16.dp))
-                
-                Column(modifier = Modifier.weight(1f)) {
-                    Text(text = uiState.name, style = MaterialTheme.typography.titleMedium, fontWeight = FontWeight.Bold, maxLines = 1)
-                    Text(text = uiState.brand, style = MaterialTheme.typography.bodySmall, color = contentColor.copy(alpha = 0.7f))
-                }
-                
-                IconButton(onClick = { onEvent(CosmeticsEvent.DeleteItem(uiState.id)) }) {
-                    Icon(Icons.Default.Delete, null, tint = contentColor.copy(alpha = 0.6f))
-                }
-            }
-            
-            Spacer(modifier = Modifier.height(12.dp))
-            
-            Row(modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween, verticalAlignment = Alignment.CenterVertically) {
-                Column {
-                    Text(stringResource(R.string.applications_kocolor_features_cosmetics_cost_per_use_label), style = MaterialTheme.typography.labelSmall, color = contentColor.copy(alpha = 0.6f))
-                    Text(text = uiState.costPerUse?.let { "$%.2f".format(it) } ?: stringResource(R.string.applications_kocolor_features_cosmetics_not_available), style = MaterialTheme.typography.bodyMedium, fontWeight = FontWeight.Bold)
-                }
-                Button(
-                    onClick = { onEvent(CosmeticsEvent.UseItem(uiState.id)) },
-                    colors = ButtonDefaults.buttonColors(containerColor = contentColor.copy(alpha = 0.2f), contentColor = contentColor),
-                    shape = RoundedCornerShape(8.dp)
-                ) {
-                    Text(stringResource(R.string.applications_kocolor_features_cosmetics_use_action), fontWeight = FontWeight.Black)
                 }
             }
         }
